fixed imprecisions in Coordinate transform
git-svn-id: http://sdoatsidc.oma.be/svn/spoca@52 96c9e701-61e4-4e2c-872f-0a44b20625ab

+++ classes/SunImage.cpp
@@ -368,8 +368,8 @@ inline HPC SunImage<T>::toHPC(const HCC& hcc) const
 	{
 		double zeta =  wcs.dsun_obs - hcc.z;
 		double dist = sqrt(hcc.x * hcc.x + hcc.y * hcc.y + zeta * zeta);
-		double rx = atan2(hcc.x, zeta);
-		double ry = asin(hcc.y / dist);
+		double rx = atan2(double(hcc.x), zeta);
+		double ry = asin(double(hcc.y) / dist);
 		return HPC(rx*RADIAN2ARCSEC, ry*RADIAN2ARCSEC);
 	}
 }
